split out utility functions into separate module
