+++ .github/workflows/main.yml
@@ -19,6 +19,7 @@ jobs:
     - name: Run linter
       run: |
         flake8 view_coverage.py
+        flake8 coverage_tools.py
 
     - name: Run tests
       run: |
