+++ .gitignore
@@ -2,3 +2,4 @@ example
 .idea
 __pycache__
 venv
+.pytest_cache
