+++ .gitlab-ci.yml
@@ -5,4 +5,5 @@ test:
     - source venv/bin/activate
     - pip install -r requirements.txt
     - flake8 view_coverage.py
-    - python -m test
+    - flake8 coverage_tools.py
+    - python -m pytest test
